Convert the reset password flow to modals, to reduce rerouting of pages
